requirements.txt
art==5.2
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.21.0
pandas==1.3.0
prettytable==2.1.0
protobuf==3.17.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
threadpoolctl==2.2.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
